services/user-service/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic[email]
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.7
email-validator==2.0.0
PyJWT==2.8.0